--- a/aeron-driver/src/main/c/aeron_driver_receiver.c
+++ b/aeron-driver/src/main/c/aeron_driver_receiver.c
@@ -46,21 +46,11 @@ int aeron_driver_receiver_init(
         return -1;
     }
 
-    for (size_t i = 0; i < AERON_DRIVER_RECEIVER_NUM_RECV_BUFFERS; i++)
+    for (size_t i = 0; i < AERON_DRIVER_UDP_NUM_RECV_BUFFERS; i++)
     {
-        size_t offset = 0;
-        if (aeron_alloc_aligned(
-            (void **)&receiver->recv_buffers.buffers[i],
-            &offset,
-            AERON_DRIVER_RECEIVER_MAX_UDP_PACKET_LENGTH,
-            AERON_CACHE_LINE_LENGTH) < 0)
-        {
-            aeron_set_err_from_last_err_code("%s:%d", __FILE__, __LINE__);
-            return -1;
-        }
-
-        receiver->recv_buffers.iov[i].iov_base = receiver->recv_buffers.buffers[i] + offset;
-        receiver->recv_buffers.iov[i].iov_len = AERON_DRIVER_RECEIVER_MAX_UDP_PACKET_LENGTH;
+        uint8_t *buffer_base = receiver->recv_buffers.buffers + i *AERON_DRIVER_UDP_RECV_BUFFER_SIZE;
+        receiver->recv_buffers.iov[i].iov_base = (void *)AERON_ALIGN((uintptr_t)(buffer_base), AERON_CACHE_LINE_LENGTH);
+        receiver->recv_buffers.iov[i].iov_len = (uint32_t)context->mtu_length;
     }
 
     if (aeron_udp_channel_data_paths_init(
@@ -122,7 +112,6 @@ void aeron_driver_receiver_on_command(void *clientd, volatile void *item)
 
 int aeron_driver_receiver_do_work(void *clientd)
 {
-    struct mmsghdr mmsghdr[AERON_DRIVER_RECEIVER_NUM_RECV_BUFFERS];
     aeron_driver_receiver_t *receiver = (aeron_driver_receiver_t *)clientd;
     int64_t bytes_received = 0;
     int work_count = 0;
@@ -130,22 +119,10 @@ int aeron_driver_receiver_do_work(void *clientd)
     work_count += (int)aeron_spsc_concurrent_array_queue_drain(
         receiver->receiver_proxy.command_queue, aeron_driver_receiver_on_command, receiver, 10);
 
-    for (size_t i = 0; i < AERON_DRIVER_RECEIVER_NUM_RECV_BUFFERS; i++)
-    {
-        mmsghdr[i].msg_hdr.msg_name = &receiver->recv_buffers.addrs[i];
-        mmsghdr[i].msg_hdr.msg_namelen = sizeof(receiver->recv_buffers.addrs[i]);
-        mmsghdr[i].msg_hdr.msg_iov = &receiver->recv_buffers.iov[i];
-        mmsghdr[i].msg_hdr.msg_iovlen = 1;
-        mmsghdr[i].msg_hdr.msg_flags = 0;
-        mmsghdr[i].msg_hdr.msg_control = NULL;
-        mmsghdr[i].msg_hdr.msg_controllen = 0;
-        mmsghdr[i].msg_len = 0;
-    }
 
     int poll_result = receiver->poller_poll_func(
         &receiver->poller,
-        mmsghdr,
-        AERON_DRIVER_RECEIVER_NUM_RECV_BUFFERS,
+        &receiver->recv_buffers,
         &bytes_received,
         receiver->data_paths.recv_func,
         receiver->recvmmsg_func,
@@ -247,11 +224,6 @@ int aeron_driver_receiver_do_work(void *clientd)
 void aeron_driver_receiver_on_close(void *clientd)
 {
     aeron_driver_receiver_t *receiver = (aeron_driver_receiver_t *)clientd;
-
-    for (size_t i = 0; i < AERON_DRIVER_RECEIVER_NUM_RECV_BUFFERS; i++)
-    {
-        aeron_free(receiver->recv_buffers.buffers[i]);
-    }
 
     aeron_free(receiver->images.array);
     aeron_free(receiver->pending_setups.array);