--- a/aeron-driver/src/main/c/aeron_driver_sender.c
+++ b/aeron-driver/src/main/c/aeron_driver_sender.c
@@ -47,20 +47,10 @@ int aeron_driver_sender_init(
         return -1;
     }
 
-    for (size_t i = 0; i < AERON_DRIVER_SENDER_NUM_RECV_BUFFERS; i++)
+    for (size_t i = 0; i < AERON_DRIVER_UDP_NUM_RECV_BUFFERS; i++)
     {
-        size_t offset = 0;
-        if (aeron_alloc_aligned(
-            (void **)&sender->recv_buffers.buffers[i],
-            &offset,
-            context->mtu_length,
-            AERON_CACHE_LINE_LENGTH) < 0)
-        {
-            aeron_set_err_from_last_err_code("%s:%d", __FILE__, __LINE__);
-            return -1;
-        }
-
-        sender->recv_buffers.iov[i].iov_base = sender->recv_buffers.buffers[i] + offset;
+        uint8_t *buffer_base = sender->recv_buffers.buffers + i *AERON_DRIVER_UDP_RECV_BUFFER_SIZE;
+        sender->recv_buffers.iov[i].iov_base = (void *)AERON_ALIGN((uintptr_t)(buffer_base), AERON_CACHE_LINE_LENGTH);
         sender->recv_buffers.iov[i].iov_len = (uint32_t)context->mtu_length;
     }
 
@@ -144,24 +134,9 @@ int aeron_driver_sender_do_work(void *clientd)
         ++sender->duty_cycle_counter >= sender->duty_cycle_ratio ||
         now_ns > sender->control_poll_timeout_ns)
     {
-        struct mmsghdr mmsghdr[AERON_DRIVER_SENDER_NUM_RECV_BUFFERS];
-
-        for (size_t i = 0; i < AERON_DRIVER_SENDER_NUM_RECV_BUFFERS; i++)
-        {
-            mmsghdr[i].msg_hdr.msg_name = &sender->recv_buffers.addrs[i];
-            mmsghdr[i].msg_hdr.msg_namelen = sizeof(sender->recv_buffers.addrs[i]);
-            mmsghdr[i].msg_hdr.msg_iov = &sender->recv_buffers.iov[i];
-            mmsghdr[i].msg_hdr.msg_iovlen = 1;
-            mmsghdr[i].msg_hdr.msg_flags = 0;
-            mmsghdr[i].msg_hdr.msg_control = NULL;
-            mmsghdr[i].msg_hdr.msg_controllen = 0;
-            mmsghdr[i].msg_len = 0;
-        }
-
         poll_result = sender->poller_poll_func(
             &sender->poller,
-            mmsghdr,
-            AERON_DRIVER_SENDER_NUM_RECV_BUFFERS,
+            &sender->recv_buffers,
             &bytes_received,
             sender->data_paths.recv_func,
             sender->recvmmsg_func,
@@ -192,11 +167,6 @@ int aeron_driver_sender_do_work(void *clientd)
 void aeron_driver_sender_on_close(void *clientd)
 {
     aeron_driver_sender_t *sender = (aeron_driver_sender_t *)clientd;
-
-    for (size_t i = 0; i < AERON_DRIVER_SENDER_NUM_RECV_BUFFERS; i++)
-    {
-        aeron_free(sender->recv_buffers.buffers[i]);
-    }
 
     aeron_udp_channel_data_paths_delete(&sender->data_paths);