--- a/aeron-driver/src/main/c/aeron_driver_sender.h
+++ b/aeron-driver/src/main/c/aeron_driver_sender.h
@@ -31,8 +31,6 @@ typedef struct aeron_driver_sender_network_publication_entry_stct
 }
 aeron_driver_sender_network_publication_entry_t;
 
-#define AERON_DRIVER_SENDER_NUM_RECV_BUFFERS (2)
-
 typedef struct aeron_driver_sender_stct
 {
     aeron_driver_sender_proxy_t sender_proxy;
@@ -46,14 +44,7 @@ typedef struct aeron_driver_sender_stct
     }
     network_publications;
 
-    struct aeron_driver_sender_buffers_stct
-    {
-        uint8_t *buffers[AERON_DRIVER_SENDER_NUM_RECV_BUFFERS];
-        struct iovec iov[AERON_DRIVER_SENDER_NUM_RECV_BUFFERS];
-        struct sockaddr_storage addrs[AERON_DRIVER_SENDER_NUM_RECV_BUFFERS];
-    }
-    recv_buffers;
-
+    aeron_udp_channel_recv_buffers_t recv_buffers;
     aeron_udp_channel_data_paths_t data_paths;
 
     int64_t *total_bytes_sent_counter;