@tmontgomery 2020-05-11
Failing builds and failing LGTM. And this is not the direction we want to go. Also, the change to the way the bindings work has impacts for us in the future.