@lygstate 2020-05-11
Performance test, Win32
Before patch:
```
Execute again? (y/n): y
Pinging 1,000,000 messages of length 32 bytes
       Value   Percentile   TotalCount 1/(1-Percentile)

      19.007     0.000000            1         1.00
      24.703     0.100000       100511         1.11
      25.807     0.200000       203590         1.25
      27.407     0.300000       300485         1.43
      29.215     0.400000       403058         1.67
      31.103     0.500000       500084         2.00
      32.703     0.550000       550652         2.22
      33.823     0.600000       600149         2.50
      35.231     0.650000       651067         2.86
      36.607     0.700000       701092         3.33
      38.207     0.750000       750079         4.00
      39.103     0.775000       778213         4.44
      39.903     0.800000       801852         5.00
      41.119     0.825000       826321         5.71
      42.527     0.850000       850177         6.67
      44.703     0.875000       876113         8.00
      45.919     0.887500       887925         8.89
      46.431     0.900000       902624        10.00
      46.815     0.912500       914668        11.43
      48.031     0.925000       925010        13.33
      48.607     0.937500       939474        16.00
      48.831     0.943750       944008        17.78
      49.727     0.950000       950084        20.00
      50.527     0.956250       956409        22.86
      51.327     0.962500       962948        26.67
      52.927     0.968750       968926        32.00
      53.407     0.971875       972068        35.56
      54.431     0.975000       975099        40.00
      55.711     0.978125       978259        45.71
      57.311     0.981250       981344        53.33
      60.607     0.984375       984440        64.00
      62.719     0.985938       985963        71.11
      65.311     0.987500       987541        80.00
      68.607     0.989062       989074        91.43
      73.215     0.990625       990639       106.67
      79.423     0.992188       992189       128.00
      83.327     0.992969       992975       142.22
      87.807     0.993750       993761       160.00
      93.631     0.994531       994532       182.86
     100.607     0.995313       995315       213.33
     110.143     0.996094       996094       256.00
     116.415     0.996484       996487       284.44
     123.455     0.996875       996875       320.00
     133.503     0.997266       997267       365.71
     147.839     0.997656       997661       426.67
     167.807     0.998047       998051       512.00
     183.551     0.998242       998245       568.89
     200.191     0.998437       998439       640.00
     223.871     0.998633       998634       731.43
     260.223     0.998828       998829       853.33
     316.159     0.999023       999024      1024.00
     351.231     0.999121       999122      1137.78
     401.151     0.999219       999220      1280.00
     460.543     0.999316       999319      1462.86
     535.039     0.999414       999415      1706.67
     668.159     0.999512       999512      2048.00
     762.367     0.999561       999562      2275.56
     862.719     0.999609       999610      2560.00
    1018.879     0.999658       999659      2925.71
    1248.255     0.999707       999708      3413.33
    1566.719     0.999756       999756      4096.00
    1755.135     0.999780       999781      4551.11
    2046.975     0.999805       999805      5120.00
    2271.231     0.999829       999830      5851.43
    2701.311     0.999854       999854      6826.67
    3170.303     0.999878       999878      8192.00
    3725.311     0.999890       999891      9102.22
    4102.143     0.999902       999903     10240.00
    4607.999     0.999915       999915     11702.86
    5488.639     0.999927       999927     13653.33
    6987.775     0.999939       999939     16384.00
    7544.831     0.999945       999946     18204.44
    8011.775     0.999951       999952     20480.00
    9322.495     0.999957       999958     23405.71
   11051.007     0.999963       999964     27306.67
   12197.887     0.999969       999970     32768.00
   12812.287     0.999973       999973     36408.89
   14262.271     0.999976       999976     40960.00
   15130.623     0.999979       999979     46811.43
   15654.911     0.999982       999982     54613.33
   17301.503     0.999985       999985     65536.00
   18661.375     0.999986       999987     72817.78
   18939.903     0.999988       999988     81920.00
   20004.863     0.999989       999990     93622.86
   20086.783     0.999991       999992    109226.67
   20348.927     0.999992       999993    131072.00
   22085.631     0.999993       999994    145635.56
   22085.631     0.999994       999994    163840.00
   25739.263     0.999995       999995    187245.71
   26918.911     0.999995       999996    218453.33
   28950.527     0.999996       999997    262144.00
   28950.527     0.999997       999997    291271.11
   28950.527     0.999997       999997    327680.00
   32227.327     0.999997       999998    374491.43
   32227.327     0.999998       999998    436906.67
   32800.767     0.999998       999999    524288.00
   32800.767     0.999998       999999    582542.22
   32800.767     0.999998       999999    655360.00
   32800.767     0.999999       999999    748982.86
   32800.767     0.999999       999999    873813.33
   36274.175     0.999999      1000000   1048576.00
   36274.175     1.000000      1000000          inf
#[Mean    =       35.726, StdDeviation   =      131.794]
#[Max     =    36274.175, Total count    =      1000000]
#[Buckets =           24, SubBuckets     =         2048]
Throughput of 27,888.547836 RTTs/sec
```

After patch:
```
Execute again? (y/n): y
Pinging 1,000,000 messages of length 32 bytes
       Value   Percentile   TotalCount 1/(1-Percentile)

      18.111     0.000000            1         1.00
      24.415     0.100000       102719         1.11
      25.215     0.200000       203757         1.25
      26.415     0.300000       301695         1.43
      27.903     0.400000       405703         1.67
      29.215     0.500000       501570         2.00
      30.111     0.550000       552753         2.22
      31.311     0.600000       604213         2.50
      32.511     0.650000       650136         2.86
      33.727     0.700000       700824         3.33
      34.719     0.750000       750068         4.00
      35.615     0.775000       775683         4.44
      36.703     0.800000       800614         5.00
      37.503     0.825000       825651         5.71
      38.719     0.850000       852142         6.67
      40.319     0.875000       875617         8.00
      41.727     0.887500       888408         8.89
      42.303     0.900000       900599        10.00
      43.807     0.912500       913890        11.43
      44.415     0.925000       925831        13.33
      46.207     0.937500       938418        16.00
      46.623     0.943750       943986        17.78
      47.103     0.950000       950630        20.00
      48.223     0.956250       957012        22.86
      48.607     0.962500       963282        26.67
      49.023     0.968750       970199        32.00
      49.215     0.971875       972769        35.56
      49.631     0.975000       975346        40.00
      50.527     0.978125       978251        45.71
      51.231     0.981250       981326        53.33
      52.703     0.984375       984402        64.00
      53.727     0.985938       986067        71.11
      55.103     0.987500       987584        80.00
      57.215     0.989062       989075        91.43
      59.903     0.990625       990633       106.67
      63.327     0.992188       992217       128.00
      65.311     0.992969       992981       142.22
      67.903     0.993750       993778       160.00
      71.039     0.994531       994543       182.86
      74.751     0.995313       995318       213.33
      80.063     0.996094       996107       256.00
      82.943     0.996484       996495       284.44
      86.335     0.996875       996880       320.00
      90.111     0.997266       997270       365.71
      95.039     0.997656       997665       426.67
     101.439     0.998047       998048       512.00
     105.407     0.998242       998245       568.89
     110.463     0.998437       998440       640.00
     115.519     0.998633       998633       731.43
     122.815     0.998828       998829       853.33
     132.223     0.999023       999027      1024.00
     138.111     0.999121       999122      1137.78
     146.303     0.999219       999223      1280.00
     153.599     0.999316       999317      1462.86
     164.351     0.999414       999415      1706.67
     182.527     0.999512       999514      2048.00
     195.967     0.999561       999561      2275.56
     212.223     0.999609       999610      2560.00
     230.911     0.999658       999659      2925.71
     248.447     0.999707       999708      3413.33
     277.759     0.999756       999756      4096.00
     297.215     0.999780       999781      4551.11
     317.951     0.999805       999805      5120.00
     345.343     0.999829       999830      5851.43
     377.343     0.999854       999854      6826.67
     423.167     0.999878       999878      8192.00
     474.367     0.999890       999891      9102.22
     529.407     0.999902       999903     10240.00
     616.447     0.999915       999915     11702.86
     724.479     0.999927       999927     13653.33
     833.535     0.999939       999939     16384.00
     935.423     0.999945       999946     18204.44
    1020.415     0.999951       999952     20480.00
    1093.631     0.999957       999958     23405.71
    1356.799     0.999963       999964     27306.67
    1457.151     0.999969       999970     32768.00
    1593.343     0.999973       999973     36408.89
    1659.903     0.999976       999976     40960.00
    1839.103     0.999979       999979     46811.43
    2113.535     0.999982       999982     54613.33
    2387.967     0.999985       999985     65536.00
    2580.479     0.999986       999987     72817.78
    2588.671     0.999988       999988     81920.00
    2756.607     0.999989       999990     93622.86
    2912.255     0.999991       999991    109226.67
    3600.383     0.999992       999993    131072.00
    3756.031     0.999993       999995    145635.56
    3756.031     0.999994       999995    163840.00
    3756.031     0.999995       999995    187245.71
    4624.383     0.999995       999996    218453.33
    4726.783     0.999996       999997    262144.00
    4726.783     0.999997       999997    291271.11
    4726.783     0.999997       999997    327680.00
    5124.095     0.999997       999998    374491.43
    5124.095     0.999998       999998    436906.67
    6250.495     0.999998       999999    524288.00
    6250.495     0.999998       999999    582542.22
    6250.495     0.999998       999999    655360.00
    6250.495     0.999999       999999    748982.86
    6250.495     0.999999       999999    873813.33
    6828.031     0.999999      1000000   1048576.00
    6828.031     1.000000      1000000          inf
#[Mean    =       31.888, StdDeviation   =       21.216]
#[Max     =     6828.031, Total count    =      1000000]
#[Buckets =           24, SubBuckets     =         2048]
Throughput of 31,241.724360 RTTs/sec

```